feat: added button action & clean up
